Adam dodaje domyślną stronę
