+++ Program/PodpisBio/PodpisBio.csproj
@@ -106,6 +106,12 @@
     <Compile Include="ShowSignatures.xaml.cs">
       <DependentUpon>ShowSignatures.xaml</DependentUpon>
     </Compile>
+    <Compile Include="View\Pages\DefaultPage.xaml.cs">
+      <DependentUpon>DefaultPage.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="View\Pages\SettingsPage.xaml.cs">
+      <DependentUpon>SettingsPage.xaml</DependentUpon>
+    </Compile>
     <Compile Include="View\Pages\ChartsPage.xaml.cs">
       <DependentUpon>ChartsPage.xaml</DependentUpon>
     </Compile>
@@ -163,6 +169,14 @@
     <Page Include="ShowSignatures.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="View\Pages\DefaultPage.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Include="View\Pages\SettingsPage.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
     </Page>
     <Page Include="View\Pages\ChartsPage.xaml">
       <SubType>Designer</SubType>
