+++ Program/PodpisBio/View/Pages/SettingsPage.xaml
@@ -1,0 +1,15 @@
+﻿<Page
+    x:Class="PodpisBio.Src.SettingsPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:PodpisBio.View"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+        <Border BorderBrush="{x:Null}" Height="50">
+            <TextBlock TextWrapping="Wrap"  Text="Strona z ustawieniami" VerticalAlignment="Center" HorizontalAlignment="Center"/>
+        </Border>
+    </Grid>
+</Page>
